// packages/cli/src/ui/types.d.ts
export declare enum MessageType {
    INFO = "info",
    ERROR = "error",
    USER = "user",
    ABOUT = "about",
    STATS = "stats",
    MODEL_STATS = "model_stats",
    TOOL_STATS = "tool_stats",
    QUIT = "quit",
    GEMINI = "gemini",
    COMPRESSION = "compression"
}